+++ src/daemon/gnet/pool.go
@@ -213,7 +213,7 @@ func (pool *ConnectionPool) Run() error {
 
 	// start the connection accept loop
 	addr := fmt.Sprintf("%s:%v", pool.Config.Address, pool.Config.Port)
-	logger.Info("Listening for connections on %s...", addr)
+	logger.Infof("Listening for connections on %s...", addr)
 
 	ln, err := net.Listen("tcp", addr)
 	if err != nil {
@@ -239,7 +239,7 @@ loop:
 				break loop
 			default:
 				// without the default case the select will block.
-				logger.Error("%v", err)
+				logger.Error(err)
 				continue
 			}
 		}
@@ -268,7 +268,7 @@ func (pool *ConnectionPool) processStrand() {
 			return
 		case req := <-pool.reqC:
 			if err := req.Func(); err != nil {
-				logger.Error("req.Func %s failed: %v", req.Name, err)
+				logger.Errorf("req.Func %s failed: %v", req.Name, err)
 			}
 		}
 	}
@@ -325,22 +325,22 @@ func (pool *ConnectionPool) ListeningAddress() (net.Addr, error) {
 
 // Creates a Connection and begins its read and write loop
 func (pool *ConnectionPool) handleConnection(conn net.Conn, solicited bool) {
-	defer logger.Debug("connection %s closed", conn.RemoteAddr())
+	defer logger.Debugf("connection %s closed", conn.RemoteAddr())
 	addr := conn.RemoteAddr().String()
 	exist, err := pool.IsConnExist(addr)
 	if err != nil {
-		logger.Error("%v", err)
+		logger.Error(err)
 		return
 	}
 
 	if exist {
-		logger.Error("Connection %s already exists", addr)
+		logger.Errorf("Connection %s already exists", addr)
 		return
 	}
 
 	c, err := pool.NewConnection(conn, solicited)
 	if err != nil {
-		logger.Error("Create connection failed: %v", err)
+		logger.Errorf("Create connection failed: %v", err)
 		return
 	}
 
@@ -393,11 +393,11 @@ func (pool *ConnectionPool) handleConnection(conn net.Conn, solicited bool) {
 	select {
 	case <-pool.quit:
 		if err := conn.Close(); err != nil {
-			logger.Error("conn.Close() error: %v", err)
+			logger.Errorf("conn.Close() error: %v", err)
 		}
 	case err = <-errC:
 		if err := pool.Disconnect(c.Addr(), err); err != nil {
-			logger.Error("Disconnect failed: %v", err)
+			logger.Errorf("Disconnect failed: %v", err)
 		}
 	}
 	close(qc)
@@ -522,7 +522,7 @@ func decodeData(buf *bytes.Buffer, maxMsgLength int) ([][]byte, error) {
 		tmpLength := uint32(0)
 		encoder.DeserializeAtomic(prefix, &tmpLength)
 		length := int(tmpLength)
-		// logger.Debug("Length is %d", length)
+		// logger.Debugf("Length is %d", length)
 		// Disconnect if we received an invalid length.
 		if length < messagePrefixLength ||
 			length > maxMsgLength {
@@ -607,7 +607,7 @@ func (pool *ConnectionPool) Connect(address string) error {
 		return nil
 	}
 
-	logger.Debug("Making TCP Connection to %s", address)
+	logger.Debugf("Making TCP Connection to %s", address)
 	conn, err := net.DialTimeout("tcp", address, pool.Config.DialTimeout)
 	if err != nil {
 		return err
@@ -670,7 +670,7 @@ func (pool *ConnectionPool) Size() (l int, err error) {
 // SendResults channel.
 func (pool *ConnectionPool) SendMessage(addr string, msg Message) error {
 	if pool.Config.DebugPrint {
-		logger.Debug("Send, Msg Type: %s", reflect.TypeOf(msg))
+		logger.Debugf("Send, Msg Type: %s", reflect.TypeOf(msg))
 	}
 
 	return pool.strand("SendMessage", func() error {
@@ -688,7 +688,7 @@ func (pool *ConnectionPool) SendMessage(addr string, msg Message) error {
 // BroadcastMessage sends a Message to all connections in the Pool.
 func (pool *ConnectionPool) BroadcastMessage(msg Message) error {
 	if pool.Config.DebugPrint {
-		logger.Debug("Broadcast, Msg Type: %s", reflect.TypeOf(msg))
+		logger.Debugf("Broadcast, Msg Type: %s", reflect.TypeOf(msg))
 	}
 
 	fullWriteQueue := []string{}