+++ src/gui/gatewayer_mock_test.go
@@ -720,6 +720,24 @@ func (m *GatewayerMock) InjectBroadcastTransaction(p0 coin.Transaction) error {
 
 }
 
+// IsWalletAPIDisabled mocked method
+func (m *GatewayerMock) IsWalletAPIDisabled() bool {
+
+	ret := m.Called()
+
+	var r0 bool
+	switch res := ret.Get(0).(type) {
+	case nil:
+	case bool:
+		r0 = res
+	default:
+		panic(fmt.Sprintf("unexpected type: %v", res))
+	}
+
+	return r0
+
+}
+
 // NewAddresses mocked method
 func (m *GatewayerMock) NewAddresses(p0 string, p1 uint64) ([]cipher.Address, error) {
 
@@ -816,6 +834,24 @@ func (m *GatewayerMock) Spend(p0 string, p1 uint64, p2 cipher.Address) (*coin.Tr
 	}
 
 	return r0, r1
+
+}
+
+// UnloadWallet mocked method
+func (m *GatewayerMock) UnloadWallet(p0 string) error {
+
+	ret := m.Called(p0)
+
+	var r0 error
+	switch res := ret.Get(0).(type) {
+	case nil:
+	case error:
+		r0 = res
+	default:
+		panic(fmt.Sprintf("unexpected type: %v", res))
+	}
+
+	return r0
 
 }