+++ src/gui/http.go
@@ -10,9 +10,8 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
-	"unicode"
-
 	"time"
+	"unicode"
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/daemon"
@@ -63,7 +62,7 @@ func create(host string, c Config, daemon *daemon.Daemon) (*Server, error) {
 	if err != nil {
 		return nil, err
 	}
-	logger.Info("Web resources directory: %s", appLoc)
+	logger.Infof("Web resources directory: %s", appLoc)
 
 	csrfStore := &CSRFStore{
 		Enabled: !c.DisableCSRF,
@@ -126,8 +125,8 @@ func CreateHTTPS(host string, c Config, daemon *daemon.Daemon, certFile, keyFile
 		return nil, err
 	}
 
-	logger.Info("Using %s for the certificate", certFile)
-	logger.Info("Using %s for the key", keyFile)
+	logger.Infof("Using %s for the certificate", certFile)
+	logger.Infof("Using %s for the key", keyFile)
 
 	cert, err := tls.LoadX509KeyPair(certFile, keyFile)
 	if err != nil {
@@ -146,7 +145,7 @@ func CreateHTTPS(host string, c Config, daemon *daemon.Daemon, certFile, keyFile
 
 // Serve serves the web interface on the configured host
 func (s *Server) Serve() error {
-	logger.Info("Starting web interface on %s", s.listener.Addr())
+	logger.Infof("Starting web interface on %s", s.listener.Addr())
 	defer logger.Info("Web interface closed")
 	defer close(s.done)
 
@@ -166,6 +165,29 @@ func (s *Server) Shutdown() {
 	<-s.done
 }
 
+func ElapseHandler(handler http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		lrw := NewWrappedResponseWriter(w)
+		start := time.Now()
+		handler.ServeHTTP(lrw, r)
+		logger.Infof("%v %s %s %v", lrw.statusCode, r.Method, r.URL.Path, time.Since(start))
+	})
+}
+
+type wrappedResponseWriter struct {
+	http.ResponseWriter
+	statusCode int
+}
+
+func NewWrappedResponseWriter(w http.ResponseWriter) *wrappedResponseWriter {
+	return &wrappedResponseWriter{w, http.StatusOK}
+}
+
+func (lrw *wrappedResponseWriter) WriteHeader(code int) {
+	lrw.statusCode = code
+	lrw.ResponseWriter.WriteHeader(code)
+}
+
 // newServerMux creates an http.ServeMux with handlers registered
 func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.ServeMux {
 	mux := http.NewServeMux()
@@ -177,6 +199,8 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	}
 
 	webHandler := func(endpoint string, handler http.Handler) {
+
+		handler = ElapseHandler(handler)
 		handler = CSRFCheck(csrfStore, handler)
 		handler = headerCheck(c.host, handler)
 		mux.Handle(endpoint, handler)
@@ -236,8 +260,7 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	// POST arguments:
 	//  id: Wallet ID
 	//  coins: Number of coins to spend
-	//  hours: Number of hours to spends
-	//  fee: Number of hours to use as fee, on top of the default fee.
+	//  dst: Destination address
 	//  Returns total amount spent if successful, otherwise error describing
 	//  failure status.
 	webHandler("/wallet/spend", walletSpendHandler(gateway))
@@ -248,19 +271,27 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	webHandler("/wallet/transactions", walletTransactionsHandler(gateway))
 
 	// Update wallet label
-	//      POST Arguments:
-	//          id: wallet id
-	//          label: wallet label
+	// POST Arguments:
+	//     id: wallet id
+	//     label: wallet label
 	webHandler("/wallet/update", walletUpdateHandler(gateway))
 
 	// Returns all loaded wallets
 	// returns sensitive information
 	webHandler("/wallets", walletsHandler(gateway))
 
+	// Returns wallets directory path
 	webHandler("/wallets/folderName", getWalletFolder(gateway))
 
-	// generate wallet seed
+	// Generate wallet seed
+	// GET Arguments:
+	//     entropy: entropy bitsize.
 	webHandler("/wallet/newSeed", newWalletSeed(gateway))
+
+	// unload wallet
+	// POST Argument:
+	//         id: wallet id
+	webHandler("/wallet/unload", walletUnloadHandler(gateway))
 
 	// Blockchain interface
 
@@ -301,7 +332,7 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	// get raw tx by txid.
 	webHandler("/rawtx", getRawTx(gateway))
 
-	// UxOUt api handler
+	// UxOut api handler
 
 	// get uxout by id.
 	webHandler("/uxout", getUxOutByID(gateway))
@@ -327,7 +358,7 @@ func newIndexHandler(appLoc string) http.HandlerFunc {
 	// Serves the main page
 	return func(w http.ResponseWriter, r *http.Request) {
 		page := filepath.Join(appLoc, indexPage)
-		logger.Debug("Serving index page: %s", page)
+		logger.Debugf("Serving index page: %s", page)
 		if r.URL.Path == "/" {
 			http.ServeFile(w, r, page)
 		} else {
@@ -407,12 +438,12 @@ func getOutputsHandler(gateway Gatewayer) http.HandlerFunc {
 
 		outs, err := gateway.GetUnspentOutputs(filters...)
 		if err != nil {
-			logger.Error("get unspent outputs failed: %v", err)
+			logger.Errorf("get unspent outputs failed: %v", err)
 			wh.Error500(w)
 			return
 		}
 
-		wh.SendOr404(w, outs)
+		wh.SendJSONOr500(logger, w, outs)
 	}
 }
 
@@ -439,7 +470,7 @@ func getBalanceHandler(gateway Gatewayer) http.HandlerFunc {
 		bals, err := gateway.GetBalanceOfAddrs(addrs)
 		if err != nil {
 			errMsg := fmt.Sprintf("Get balance failed: %v", err)
-			logger.Error("%s", errMsg)
+			logger.Error(errMsg)
 			wh.Error500Msg(w, errMsg)
 			return
 		}
@@ -460,7 +491,7 @@ func getBalanceHandler(gateway Gatewayer) http.HandlerFunc {
 			}
 		}
 
-		wh.SendOr404(w, balance)
+		wh.SendJSONOr500(logger, w, balance)
 	}
 }
 
@@ -471,7 +502,7 @@ func versionHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
-		wh.SendOr404(w, gateway.GetBuildInfo())
+		wh.SendJSONOr500(logger, w, gateway.GetBuildInfo())
 	}
 }
 
@@ -525,7 +556,7 @@ func getLogsHandler(logbuf *bytes.Buffer) http.HandlerFunc {
 			}
 
 			if len(logs) >= linenum {
-				logger.Debug("logs size %d,total size:%d", len(logs), len(logList))
+				logger.Debugf("logs size %d,total size:%d", len(logs), len(logList))
 				break
 			}
 			log := attrActualLog(logInfo)
@@ -535,6 +566,6 @@ func getLogsHandler(logbuf *bytes.Buffer) http.HandlerFunc {
 
 		}
 
-		wh.SendOr404(w, logs)
+		wh.SendJSONOr500(logger, w, logs)
 	}
 }