+++ src/visor/blockchain.go
@@ -393,7 +393,7 @@ func (bc Blockchain) GetBlocks(start, end uint64) []coin.SignedBlock {
 	for i := start; i <= end; i++ {
 		b, err := bc.store.GetBlockBySeq(i)
 		if err != nil {
-			logger.Error("%v", err)
+			logger.Error(err)
 			return []coin.SignedBlock{}
 		}