+++ CHANGELOG.md
@@ -29,9 +29,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `/richlist` API method, returns top n address balances
 - `/addresscount` API method, returns the number of addresses that have any amount of coins
 - `/transactions` API method, returns transactions of addresses
+- `/wallet/unload` API method, removes the wallet of given id from wallet services
 
 ### Fixed
 
+- #1021, remove `SendOr404` and `SendOr500` as they do not work properly due to typed nils
 - Add Read, Write and Idle timeouts to the HTTP listener, preventing file descriptor leaks
 - Support absolute and relative paths for `-data-dir` option
 - Prevent creating transactions whose size exceeds the maximum block size
@@ -40,6 +42,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - #885, Add `Host` header check to localhost HTTP interfaces to prevent DNS rebinding attacks
 - #896, Add CSRF check to wallet API
 - Fix base58 address parsing, which allowed leading invalid characters and treated unknown characters as a '1'
+- Fix occasional error which causes blockchain progress not to be shown in front-end
 
 ### Changed